pyproject.toml
[project]
name = "calc-bench"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "openai>=1.102.0",
    "openai-agents>=0.2.10",
    "pytest>=8.4.1",
    "python-dotenv>=1.1.1",
]

[tool.pytest.ini_options]
markers = ["integration: marks integration tests"]
